backend/requirements.txt
fastapi
uvicorn
langchain_openai
browser-use
python-dotenv
couchbase
